Change to the make_rz_plots.py (and other versions) scripts to fix a bug with calculating the radius for Polar Conversion

+++ python/make_rz_plots.py
@@ -302,7 +302,7 @@ for (FILE,fff) in zip(UNPLOTTED_LIST,np.array(range(UNPLOTTED_LIST.size))):
 			y_sr = lat_sr*111.1e3
 
 			#Define the polar coordinates needed
-			r = np.linspace(0,rmax,(rmax//resolution+1))
+			r = np.linspace(0,rmax,(np.int(rmax//resolution)+1))
 			#r = np.linspace(0,600,201)
 			#print(r)
 			#sys.exit()
